requirements.txt
asgiref==3.2.7
autopep8==1.5.2
colorama==0.4.3
dj-database-url==0.5.0
Django==3.0.5
django-crispy-forms==1.9.0
gunicorn==20.0.4
mysqlclient==1.4.6
Pillow==7.1.2
psycopg2==2.8.5
pycodestyle==2.5.0
pytz==2019.3
sqlparse==0.3.1
whitenoise==5.0.1
